Work & Chill Windows App
This is an App created in c# using Visual Studio .Net. It is a little app that plays a techno radio station and displays a random image from Google images. The window frame is in c#, but the random images and player are in PHP + JavaScript. The PHP files are running on my web server, and it is easier for me or anyone to change the content of these files.